+++ core/src/network/peer.rs
@@ -23,16 +23,17 @@ use crate::{
     Transport,
     StreamMuxer,
     connection::{
+        Connected,
+        ConnectedPoint,
         ConnectionHandler,
         ConnectionInfo,
-        Connected,
         Connection,
         ConnectionId,
         ConnectionLimit,
-        DialingConnection,
         EstablishedConnection,
         EstablishedConnectionIter,
         IntoConnectionHandler,
+        PendingConnection,
         Substream,
     },
 };
@@ -510,6 +511,74 @@ where
                 network: self.network,
                 peer_id: self.peer_id,
             })
+    }
+}
+
+/// Attempt to reach a peer.
+#[derive(Debug, Clone)]
+pub(super) struct DialingAttempt {
+    /// Identifier for the reach attempt.
+    pub(super) id: ConnectionId,
+    /// Multiaddr currently being attempted.
+    pub(super) current: Multiaddr,
+    /// Multiaddresses to attempt if the current one fails.
+    pub(super) next: Vec<Multiaddr>,
+}
+
+/// A `DialingConnection` is a [`PendingConnection`] where the local peer
+/// has the role of the dialer (i.e. initiator) and the (expected) remote
+/// peer ID is known.
+pub struct DialingConnection<'a, TInEvent, TConnInfo, TPeerId> {
+    peer_id: &'a TPeerId,
+    inner: PendingConnection<'a, TInEvent, TConnInfo, TPeerId>,
+    dialing: hash_map::OccupiedEntry<'a, TPeerId, DialingAttempt>,
+}
+
+impl<'a, TInEvent, TConnInfo, TPeerId>
+    DialingConnection<'a, TInEvent, TConnInfo, TPeerId>
+{
+    /// Returns the local connection ID.
+    pub fn id(&self) -> ConnectionId {
+        self.inner.id()
+    }
+
+    /// Returns the (expected) peer ID of the ongoing connection attempt.
+    pub fn peer_id(&self) -> &TPeerId {
+        self.peer_id
+    }
+
+    /// Returns information about this endpoint of the connection attempt.
+    pub fn endpoint(&self) -> &ConnectedPoint {
+        self.inner.endpoint()
+    }
+
+    /// Aborts the connection attempt.
+    pub fn abort(self)
+    where
+        TPeerId: Eq + Hash + Clone,
+    {
+        self.dialing.remove();
+        self.inner.abort();
+    }
+
+    /// Adds new candidate addresses to the end of the addresses used
+    /// in the ongoing dialing process.
+    ///
+    /// Duplicates are ignored.
+    pub fn add_addresses(&mut self, addrs: impl IntoIterator<Item = Multiaddr>) {
+        for addr in addrs {
+            self.add_address(addr);
+        }
+    }
+
+    /// Adds an address to the end of the addresses used in the ongoing
+    /// dialing process.
+    ///
+    /// Duplicates are ignored.
+    pub fn add_address(&mut self, addr: Multiaddr) {
+        if self.dialing.get().next.iter().all(|a| a != &addr) {
+            self.dialing.get_mut().next.push(addr);
+        }
     }
 }
